// frontend/src/enums.ts
export enum UserRole {
  CHURCH = 'church',
  SEMINARY = 'seminary',
  PRIEST = 'priest',
  SEMINARIST = 'seminarist',
  PILGRIM = 'pilgrim',
}

export enum HttpStatus {
  SUCCESS = 200,
  CREATED = 201,
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  NOT_FOUND = 404,
  INTERNAL_SERVER_ERROR = 500,
}